refactor(backend): split torch_backend.py into category mixin modules (Phase 3.1)
torch_backend.py (2211 lines) mixed capability flags, creation,
indexing, reductions, passthrough overrides, linalg, interpolation,
random, and misc operations in one file. Split into same-directory
_torch_*.py mixin modules by category, and moved the GradMode/_Config
helpers into _torch_config.py to avoid a circular import from the
capabilities mixin. TorchBackend now composes the mixins via multiple
inheritance with identical MRO/behavior. base.py (the AbstractBackend
ABC) is untouched, per spec. Pure file reorganization -- no method moved
logic, only location. Full test suite (7258 tests) and golden-value
regression snapshots pass unchanged at both backends.
